Ready for first release version

+++ GBCLV3/Utils/ColorUtil.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Media;
 
 namespace GBCLV3.Utils
@@ -15,7 +11,7 @@ namespace GBCLV3.Utils
             float U = -0.147f * color.ScR - 0.289f * color.ScG + 0.436f * color.ScB;
             float V = 0.615f * color.ScR - 0.515f * color.ScG - 0.100f * color.ScB;
 
-            Y = (float)Math.Pow((double)Y, gamma);
+            Y = (float)Math.Pow(Y, gamma);
 
             color.ScR = Y + 1.140f * V;
             color.ScG = Y - 0.395f * U - 0.581f * V;
